Splitting the index.d.ts file into smaller files
